refactor: move components and cleanup
